3 BLOKER onarımı: sessiz i32 fallback'i elendi (ön-merge denetimi bulgusu) [D-295]
NASIL BULUNDU: main'e merge ÖNCESİ adversarial denetim (6 boyut paralel inceleme +
her bulgu için 3 bağımsız çürütme lensi). 26 aday -> 15 onaylanan, 3 BLOKER.
Üçü de bağımsız derleyip-çalıştırarak teyit edildi. MERGE DURDURULDU, önce onarıldı.

ORTAK KÖK: codegen T'yi kurtaramayınca sessizce "i32" varsayıyordu. D-291→D-294
taşıyıcıyı i64'e genişletti ama fallback'leri güncellemedi -> i64 taşıyıcı kırpıldı.

  # | program                                        | ÖNCE          | SONRA
  1 | değişken c = görev_başlat(|| tam64); birleştir | exit 99 (!)   | exit 1  ✓
 1b | aynısı, T=metin                                | SEGFAULT      | exit 5  ✓
  2 | kanal<tam64> + --llvm (tip kontrolü atlanır)   | exit 1  (!)   | exit 42 ✓
  3 | değişken f = || "selam"; metin_uzunluk(f())    | SEGFAULT      | exit 5  ✓

EN AĞIR OLGU: aynı programlar origin/main'de "error: '%0' defined with type 'ptr'
but expected 'i32'" ile DERLENMİYORDU (ölçüldü). Yani D-291→D-294 hata modunu
LOUD -> SILENT çeviriyordu — projenin kendi ilkesinin birebir ihlali. D-292'nin
"sessiz-yanlış-cevap kapatıldı" başlığı bu yüzden YANLIŞTI.

ONARIMLAR:
1. görev_birleştir fallback i32->i64. Kesirli T zaten reddedildiği için (D-294)
   i64 güvenli: hedeflerde tamsayı ve işaretçi AYNI yazmaçta döner.
2. Closure çağrı yeri fallback i32->i64. KALAN AÇIK (dürüstçe): kesirli dönen
   annotasyonsuz closure + beklenen-yok hâlâ yanlış yazmaçtan okur; ama o vaka
   bugün de aşağı akışta GÜRÜLTÜLÜ LLVM reddi alıyor (ölçüldü) — i64 onu
   sessizleştirmiyor. Tam çözüm ayrı iş.
3. Kanal runtime int32_t->int64_t (host + bare-metal, AYNI ABI) + codegen
   T-farkında gönder/al (kanal_ic_ir; sext/ptrtoint <-> trunc/inttoptr).

   NEDEN "gürültülü yap" DEĞİL de "sınıfı yok et": ilk planım geniş T'de doğal tip
   geçirip declare uyuşmazlığıyla LLVM'i reddettirmekti. ÖLÇTÜM: LLVM imza
   uyuşmazlığını SESSİZCE KABUL EDİYOR (opt -passes=verify ve llvm-as exit 0)
   -> plan geçersizdi. Ölçmeseydim blokeri kapattığımı sanacaktım.

KAPSAM GENİŞLEDİ: D-292'nin 32-bit kısıtı KALKTI. kanal<tam64>, kanal<metin>,
kanal<Dizi<T>> artık GERÇEKTEN çalışıyor. Kalan kısıt kesirli T (DRF006); katmanlı
savunma kanal-float için de ölçüldü: --llvm atlasa bile emisyon `sext double -> i64`
üretir, GEÇERSİZDİR -> LLVM gürültülü reddeder.

TEST BEKLENTİSİ DEĞİŞTİ (dürüstçe): D4/D6/D45 artık DRF006 değil "0 hata" bekliyor;
yeni D50 kesirli reddini kilitler.

SAHTE-TEST DOĞRULAMASI: fallback'i i32'ye geri alan SABOTAJ ile yeni testler
[253]/[254]/[255] KIRMIZIYA döndü -> gerçekten kilitliyorlar. [256]/[257] kanal
runtime genişletmesine bağlı; taban ölçümü (fix öncesi exit 1 / LLVM reddi) onların
da ayırt edici olduğunu gösteriyor.

KANIT: test_llvm 258/258 (+6, hepsi ANNOTASYONSUZ biçimleri kullanır — kırık olan
idiomatik biçimdi). test_drf 50/50. test_gorev_rt 13/13. Regresyon: bare-metal
calistir_kanal_test_arm geçti (toplam=55), codegen parite 76/76, SELF-HOST FIXPOINT ✓
(33371 satır), test_tumu "Tum testler gecti!".

Ayrıca ölçüldü: dar/işaretsiz T turları doğru (kanal<tam16> 300, kanal<dtam8> 200,
kanal<tam8> -1, görev<tam16> 300 — hepsi kayıpsız).

SÜREÇ DERSİ: taşıyıcı genişliğini değiştirince TÜM fallback'leri denetle. LLVM
declare imza uyuşmazlığı bir savunma mekanizması DEĞİLDİR.

+++ runtime/kdl_kanal.c
@@ -25,7 +25,7 @@
 #define KDL_KANAL_HAVUZ  4   /* eşzamanlı kanal sayısı */
 
 struct KdlKanal {
-    volatile int buf[KDL_KANAL_KAP];
+    volatile int64_t buf[KDL_KANAL_KAP];   /* D-295: host ile AYNI genislik */
     volatile int bas;   /* okuma (tüketici) indeksi */
     volatile int son;   /* yazma (üretici) indeksi */
 };
@@ -53,15 +53,15 @@ static int kdl_kanal_bos(struct KdlKanal *k) {
     return k->son == k->bas;
 }
 
-void kdl_kanal_gonder(KdlKanal *k, int deger) {
+void kdl_kanal_gonder(KdlKanal *k, int64_t deger) {
     while (kdl_kanal_dolu(k)) { __asm__ volatile("" ::: "memory"); }
     k->buf[k->son] = deger;
     k->son = (k->son + 1) % KDL_KANAL_KAP;
 }
 
-int kdl_kanal_al(KdlKanal *k) {
+int64_t kdl_kanal_al(KdlKanal *k) {
     while (kdl_kanal_bos(k)) { __asm__ volatile("" ::: "memory"); }
-    int v = k->buf[k->bas];
+    int64_t v = k->buf[k->bas];
     k->bas = (k->bas + 1) % KDL_KANAL_KAP;
     return v;
 }
